⤵️Move style files to separate folder
