+++ client/src/Components/Sites/Home.tsx
@@ -21,7 +21,7 @@ import { FormattedMessage } from "react-intl";
 import { Link } from "react-router-dom";
 import { Button, Container } from "@mui/material";
 import RowingOutlined from "@mui/icons-material/RowingOutlined";
-import { useDefaultStyles } from "../../css/Style";
+import { useDefaultStyles } from "../../style/Style";
 
 export function Home(): JSX.Element {
     const classes = useDefaultStyles();
