+++ client/src/Components/Sites/Settings.tsx
@@ -21,7 +21,7 @@ import { FormattedMessage } from "react-intl";
 import FormControl from "@mui/material/FormControl";
 import { Container, InputLabel, MenuItem, Select, SelectChangeEvent } from "@mui/material";
 
-import { useDefaultStyles } from "../../css/Style";
+import { useDefaultStyles } from "../../style/Style";
 import { useLanguage } from "../../state/actions/languageActions";
 
 type Language = { code: string; name: string };
