+++ client/src/style/LeaderboardStyle.ts
